pyproject.toml
[project]
name = "trendfilterpy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Erik Dains", email = "eadains@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "clarabel>=0.9.0",
    "cvxpy-base>=1.6.0",
    "numpy>=2.2.1",
    "scikit-learn>=1.6.0",
    "scipy>=1.14.1",
]

[project.scripts]
trendfilterpy = "trendfilterpy:main"

[project.optional-dependencies]
tests = [
    "polars>=1.18.0",
    "pytest>=8.3.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
# Defaults plus pydocstyle, flake8-bugbear, flake8-simplify, annotations, and import sorting
select = ["E4", "E7", "E9", "F", "D", "B", "SIM", "ANN", "I"]
# Ignore module docstrings
ignore = ["D100", "D104"]

[tool.ruff.format]
skip-magic-trailing-comma = true

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.0",
    "ruff>=0.8.4",
]

[tool.ruff.lint.pydocstyle]
convention = "numpy"
